fix(vectors): scan embedded secret payload values
