+++ crates/axon-vectors/Cargo.toml
@@ -10,8 +10,8 @@ license.workspace = true
 async-trait = "0.1"
 axon-api = { path = "../axon-api" }
 axon-error = { path = "../axon-error" }
+serde_json = "1"
 tokio = { version = "1", features = ["macros", "rt", "sync"] }
 uuid = { version = "1", features = ["v4", "v5", "serde"] }
 
 [dev-dependencies]
-serde_json = "1"
