+++ crates/axon-vectors/src/payload.rs
@@ -4,7 +4,8 @@ use std::collections::{BTreeMap, BTreeSet};
 use std::fmt;
 
 use axon_api::source::MetadataMap;
-use serde_json::Value;
+
+use crate::payload_redaction::{forbidden_field_name, validate_forbidden_value};
 
 pub const MODULE_NAME: &str = "payload";
 
@@ -250,118 +251,6 @@ fn validate_known_fields(
     Ok(())
 }
 
-fn forbidden_field_name(field: &str) -> bool {
-    let normalized = field.to_ascii_lowercase();
-    FORBIDDEN_FIELD_FRAGMENTS
-        .iter()
-        .any(|fragment| normalized.contains(fragment))
-}
-
-fn validate_forbidden_value(path: &str, value: &Value) -> Result<(), VectorPayloadValidationError> {
-    match value {
-        Value::String(value) if forbidden_string_value(value) => {
-            Err(VectorPayloadValidationError::ForbiddenValue {
-                field: path.to_string(),
-            })
-        }
-        Value::Array(values) => {
-            for (index, value) in values.iter().enumerate() {
-                validate_forbidden_value(&format!("{path}[{index}]"), value)?;
-            }
-            Ok(())
-        }
-        Value::Object(object) => {
-            if adapter_response_blob(object) {
-                return Err(VectorPayloadValidationError::ForbiddenValue {
-                    field: path.to_string(),
-                });
-            }
-            for (field, value) in object {
-                let child_path = format!("{path}.{field}");
-                if forbidden_field_name(field) {
-                    return Err(VectorPayloadValidationError::ForbiddenValue { field: child_path });
-                }
-                validate_forbidden_value(&child_path, value)?;
-            }
-            Ok(())
-        }
-        _ => Ok(()),
-    }
-}
-
-fn forbidden_string_value(value: &str) -> bool {
-    let normalized = value.to_ascii_lowercase();
-    FORBIDDEN_VALUE_FRAGMENTS
-        .iter()
-        .any(|fragment| normalized.contains(fragment))
-        || raw_dotenv_assignment(value)
-        || bare_secret_token(value)
-        || home_credential_path(&normalized)
-        || raw_html_blob(&normalized)
-        || normalized.contains("adapter_response")
-}
-
-fn raw_dotenv_assignment(value: &str) -> bool {
-    value.lines().any(|line| {
-        let line = line.trim();
-        let Some((key, raw_value)) = line.split_once('=') else {
-            return false;
-        };
-        let key = key.trim();
-        !key.is_empty()
-            && !raw_value.trim().is_empty()
-            && key
-                .chars()
-                .all(|ch| ch.is_ascii_uppercase() || ch.is_ascii_digit() || ch == '_')
-            && key
-                .chars()
-                .next()
-                .is_some_and(|ch| ch.is_ascii_uppercase() || ch == '_')
-    })
-}
-
-fn bare_secret_token(value: &str) -> bool {
-    let trimmed = value.trim();
-    BARE_SECRET_TOKEN_PREFIXES
-        .iter()
-        .any(|prefix| bare_secret_token_with_prefix(trimmed, prefix))
-}
-
-fn bare_secret_token_with_prefix(value: &str, prefix: &str) -> bool {
-    let Some(rest) = value.strip_prefix(prefix) else {
-        return false;
-    };
-    let secret_chars = rest
-        .chars()
-        .take_while(|ch| ch.is_ascii_alphanumeric() || matches!(ch, '_' | '-'))
-        .count();
-    secret_chars >= 20
-}
-
-fn home_credential_path(normalized: &str) -> bool {
-    normalized.contains("/home/")
-        && HOME_CREDENTIAL_PATH_FRAGMENTS
-            .iter()
-            .any(|fragment| normalized.contains(fragment))
-}
-
-fn raw_html_blob(normalized: &str) -> bool {
-    let trimmed = normalized.trim_start();
-    trimmed.starts_with("<!doctype html")
-        || trimmed.starts_with("<html")
-        || (normalized.contains("<html") && normalized.contains("</html>"))
-        || (normalized.contains("<body") && normalized.contains("</body>"))
-}
-
-fn adapter_response_blob(object: &serde_json::Map<String, Value>) -> bool {
-    let has_status = object.contains_key("status") || object.contains_key("status_code");
-    let has_headers = object.contains_key("headers");
-    let has_body = object.contains_key("body")
-        || object.contains_key("raw_body")
-        || object.contains_key("response_body");
-    has_status && has_headers && has_body
-}
-
 const REQUIRED_FIELDS: &[&str] = &[
     "payload_contract_version",
     "collection",
@@ -403,60 +292,4 @@ const SHARED_FIELDS: &[&str] = &[
     "embedding_provider",
     "embedding_profile",
     "embedded_at",
-];
-
-const FORBIDDEN_FIELD_FRAGMENTS: &[&str] = &[
-    "raw_auth",
-    "auth_header",
-    "authorization",
-    "cookie",
-    "api_key",
-    "apikey",
-    "secret",
-    "raw_env",
-    "env_value",
-    "absolute_home",
-    "home_path",
-    "raw_html",
-    "html_blob",
-    "adapter_response",
-    "response_blob",
-];
-
-const FORBIDDEN_VALUE_FRAGMENTS: &[&str] = &[
-    "authorization:",
-    "proxy-authorization:",
-    "bearer ",
-    "cookie:",
-    "set-cookie:",
-    "api_key=",
-    "apikey=",
-    "api-key:",
-    "x-api-key:",
-    "access_token=",
-    "refresh_token=",
-    "secret_key=",
-    "token=",
-];
-
-const BARE_SECRET_TOKEN_PREFIXES: &[&str] = &[
-    "sk-",
-    "sk_",
-    "sk-proj-",
-    "ghp_",
-    "github_pat_",
-    "xoxb-",
-    "xoxp-",
-    "glpat-",
-];
-
-const HOME_CREDENTIAL_PATH_FRAGMENTS: &[&str] = &[
-    "/.ssh/",
-    "/.aws/",
-    "/.gnupg/",
-    "/.config/chezmoi/key.txt",
-    "/.config/gcloud/",
-    "/.docker/config.json",
-    "/.kube/config",
-    "/.env",
 ];
