+++ crates/axon-vectors/src/payload_redaction.rs
@@ -1,0 +1,178 @@
+//! Redaction guardrails for vector payload metadata.
+
+use serde_json::Value;
+
+use crate::payload::VectorPayloadValidationError;
+
+pub(crate) fn forbidden_field_name(field: &str) -> bool {
+    let normalized = field.to_ascii_lowercase();
+    FORBIDDEN_FIELD_FRAGMENTS
+        .iter()
+        .any(|fragment| normalized.contains(fragment))
+}
+
+pub(crate) fn validate_forbidden_value(
+    path: &str,
+    value: &Value,
+) -> Result<(), VectorPayloadValidationError> {
+    match value {
+        Value::String(value) if forbidden_string_value(value) => {
+            Err(VectorPayloadValidationError::ForbiddenValue {
+                field: path.to_string(),
+            })
+        }
+        Value::Array(values) => {
+            for (index, value) in values.iter().enumerate() {
+                validate_forbidden_value(&format!("{path}[{index}]"), value)?;
+            }
+            Ok(())
+        }
+        Value::Object(object) => {
+            if adapter_response_blob(object) {
+                return Err(VectorPayloadValidationError::ForbiddenValue {
+                    field: path.to_string(),
+                });
+            }
+            for (field, value) in object {
+                let child_path = format!("{path}.{field}");
+                if forbidden_field_name(field) {
+                    return Err(VectorPayloadValidationError::ForbiddenValue { field: child_path });
+                }
+                validate_forbidden_value(&child_path, value)?;
+            }
+            Ok(())
+        }
+        _ => Ok(()),
+    }
+}
+
+fn forbidden_string_value(value: &str) -> bool {
+    let normalized = value.to_ascii_lowercase();
+    FORBIDDEN_VALUE_FRAGMENTS
+        .iter()
+        .any(|fragment| normalized.contains(fragment))
+        || raw_dotenv_assignment(value)
+        || contains_bare_secret_token(value)
+        || absolute_home_path(&normalized)
+        || raw_html_blob(&normalized)
+        || normalized.contains("adapter_response")
+}
+
+fn raw_dotenv_assignment(value: &str) -> bool {
+    value.lines().any(|line| {
+        let line = line.trim();
+        let Some((key, raw_value)) = line.split_once('=') else {
+            return false;
+        };
+        let key = key.trim();
+        !key.is_empty()
+            && !raw_value.trim().is_empty()
+            && key
+                .chars()
+                .all(|ch| ch.is_ascii_uppercase() || ch.is_ascii_digit() || ch == '_')
+            && key
+                .chars()
+                .next()
+                .is_some_and(|ch| ch.is_ascii_uppercase() || ch == '_')
+    })
+}
+
+fn contains_bare_secret_token(value: &str) -> bool {
+    BARE_SECRET_TOKEN_PREFIXES
+        .iter()
+        .any(|prefix| contains_bare_secret_token_with_prefix(value, prefix))
+}
+
+fn contains_bare_secret_token_with_prefix(value: &str, prefix: &str) -> bool {
+    let mut search_start = 0;
+    while let Some(relative_index) = value[search_start..].find(prefix) {
+        let index = search_start + relative_index;
+        let rest_start = index + prefix.len();
+        if token_start_boundary(value, index) && token_body_len(&value[rest_start..]) >= 20 {
+            return true;
+        }
+        search_start = rest_start;
+    }
+    false
+}
+
+fn token_start_boundary(value: &str, index: usize) -> bool {
+    value[..index]
+        .chars()
+        .next_back()
+        .is_none_or(|ch| !is_token_char(ch))
+}
+
+fn token_body_len(value: &str) -> usize {
+    value.chars().take_while(|ch| is_token_char(*ch)).count()
+}
+
+fn is_token_char(ch: char) -> bool {
+    ch.is_ascii_alphanumeric() || matches!(ch, '_' | '-')
+}
+
+fn absolute_home_path(normalized: &str) -> bool {
+    normalized.contains("/home/")
+}
+
+fn raw_html_blob(normalized: &str) -> bool {
+    let trimmed = normalized.trim_start();
+    trimmed.starts_with("<!doctype html")
+        || trimmed.starts_with("<html")
+        || (normalized.contains("<html") && normalized.contains("</html>"))
+        || (normalized.contains("<body") && normalized.contains("</body>"))
+}
+
+fn adapter_response_blob(object: &serde_json::Map<String, Value>) -> bool {
+    let has_status = object.contains_key("status") || object.contains_key("status_code");
+    let has_headers = object.contains_key("headers");
+    let has_body = object.contains_key("body")
+        || object.contains_key("raw_body")
+        || object.contains_key("response_body");
+    has_status && has_headers && has_body
+}
+
+const FORBIDDEN_FIELD_FRAGMENTS: &[&str] = &[
+    "raw_auth",
+    "auth_header",
+    "authorization",
+    "cookie",
+    "api_key",
+    "apikey",
+    "secret",
+    "raw_env",
+    "env_value",
+    "absolute_home",
+    "home_path",
+    "raw_html",
+    "html_blob",
+    "adapter_response",
+    "response_blob",
+];
+
+const FORBIDDEN_VALUE_FRAGMENTS: &[&str] = &[
+    "authorization:",
+    "proxy-authorization:",
+    "bearer ",
+    "cookie:",
+    "set-cookie:",
+    "api_key=",
+    "apikey=",
+    "api-key:",
+    "x-api-key:",
+    "access_token=",
+    "refresh_token=",
+    "secret_key=",
+    "token=",
+];
+
+const BARE_SECRET_TOKEN_PREFIXES: &[&str] = &[
+    "sk-proj-",
+    "github_pat_",
+    "sk-",
+    "sk_",
+    "ghp_",
+    "xoxb-",
+    "xoxp-",
+    "glpat-",
+];
