+++ crates/axon-vectors/tests/fixtures/payload/forbidden_embedded_bare_api_key_value.invalid.json
@@ -1,0 +1,25 @@
+{
+  "payload_contract_version": "2026-07-01",
+  "collection": "axon",
+  "source_family": "web",
+  "source_id": "src-web",
+  "source_generation": 3,
+  "committed_generation": 3,
+  "document_id": "doc-web",
+  "chunk_id": "chunk-web-1",
+  "chunk_locator": "https://example.com/docs#intro",
+  "source_range": { "selector": "#intro" },
+  "visibility": "public",
+  "redaction_status": "clean",
+  "job_id": "00000000-0000-0000-0000-000000000002",
+  "document_status": "ready",
+  "embedding_model": "text-embedding-test",
+  "embedding_dimensions": 1024,
+  "embedding_provider": "tei",
+  "embedding_profile": "document",
+  "embedded_at": "2026-07-01T12:00:00Z",
+  "web_title": "debug key sk-proj-abcdefghijklmnopqrstuvwxyz0123456789",
+  "web_domain": "example.com",
+  "web_status_code": 200,
+  "web_depth": 1
+}
